+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.9.0 (tbd)
 
 - Improved output when installation of packages fails (#759)
+- Improved error message when template resolution fails (#763)
 
 ## 1.8.5 (April 15, 2025)